+++ APIBlox.NetCore.EventStore/Exceptions/EventStoreNotFoundException.cs
@@ -2,33 +2,30 @@
 
 namespace APIBlox.NetCore.Exceptions
 {
-    /// <inheritdoc />
     /// <summary>
-    ///     Class DataAccessException.
+    ///     Class EventStoreNotFoundException.
     /// </summary>
-    /// <seealso cref="T:System.Exception" />
-    public class DataAccessException : Exception
+    /// <seealso cref="System.Exception" />
+    public class EventStoreNotFoundException : Exception
     {
-        /// <inheritdoc />
         /// <summary>
-        ///     Initializes a new instance of the <see cref="T:APIBlox.NetCore.Exceptions.DataAccessException" /> class.
+        ///     Initializes a new instance of the <see cref="EventStoreNotFoundException" /> class.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        public DataAccessException(string message)
+        public EventStoreNotFoundException(string message)
             : base(message)
         {
         }
 
-        /// <inheritdoc />
         /// <summary>
-        ///     Initializes a new instance of the <see cref="T:APIBlox.NetCore.Exceptions.DataAccessException" /> class.
+        ///     Initializes a new instance of the <see cref="EventStoreNotFoundException" /> class.
         /// </summary>
         /// <param name="message">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         ///     The exception that is the cause of the current exception, or a null reference (Nothing in
         ///     Visual Basic) if no inner exception is specified.
         /// </param>
-        public DataAccessException(string message, Exception innerException)
+        public EventStoreNotFoundException(string message, Exception innerException)
             : base(message, innerException)
         {
         }