+++ Examples/Event Sourcing/Examples.EventSourcing/MongoAggregate.cs
@@ -12,13 +12,16 @@ using Newtonsoft.Json;
 
 namespace Examples
 {
-    public class MongoAggregate 
+    /// <summary>
+    ///     Class MongoAggregate.
+    /// </summary>
+    public class MongoAggregate
     {
         private readonly IEventStoreService<MongoAggregate> _es;
+        private EventStreamModel _myEventStream;
 
         private readonly string _streamId;
         private readonly IDictionary<Type, MethodInfo> _whenMethods;
-        private EventStreamModel _myEventStream;
 
         /// <summary>
         ///     Initializes a new instance of the <see cref="MongoAggregate" /> class.
@@ -56,8 +59,11 @@ namespace Examples
         /// <value>Some value.</value>
         public string SomeValue { get; private set; }
 
+        /// <summary>
+        ///     Gets my version.
+        /// </summary>
+        /// <value>My version.</value>
         public long MyVersion { get; private set; }
-
 
         /// <summary>
         ///     Adds some value.
@@ -65,13 +71,13 @@ namespace Examples
         /// <param name="someValue">Some value.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>Task.</returns>
-        /// <exception cref="DataAccessException">Aggregate with stream id {_streamId}</exception>
+        /// <exception cref="EventStoreConcurrencyException">Aggregate with stream id {_streamId}</exception>
         public async Task AddSomeValue(string someValue, CancellationToken cancellationToken)
         {
             await Build(false, cancellationToken);
 
             if (!(_myEventStream is null))
-                throw new DataAccessException($"Another Aggregate with stream id {_streamId} already exists!");
+                throw new EventStoreConcurrencyException($"Another Aggregate with stream id {_streamId} already exists!");
 
             _myEventStream = new EventStreamModel();
 
@@ -89,13 +95,13 @@ namespace Examples
         /// <param name="someValue">Some value.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>Task.</returns>
-        /// <exception cref="DataAccessException">Aggregate with stream id {_streamId}</exception>
+        /// <exception cref="EventStoreNotFoundException">Aggregate with stream id {_streamId}</exception>
         public async Task UpdateSomeValue(string someValue, CancellationToken cancellationToken)
         {
             await Build(false, cancellationToken);
 
             if (_myEventStream is null)
-                throw new DataAccessException($"Another Aggregate with stream id {_streamId} not found!");
+                throw new EventStoreNotFoundException($"Another Aggregate with stream id {_streamId} not found!");
 
             // Validate and such
             SomeValue = someValue;
@@ -111,9 +117,9 @@ namespace Examples
         public async Task PublishChangesAsync(CancellationToken cancellationToken = default)
         {
             var result = await _es.WriteToEventStreamAsync(_streamId,
-                DomainEvents.Select(e => new EventModel { Data = e }).ToArray(),
-                _myEventStream.Version > 0 ? _myEventStream.Version : (long?)null,
-                cancellationToken: cancellationToken
+                DomainEvents.Select(e => new EventModel {Data = e}).ToArray(),
+                _myEventStream.Version > 0 ? _myEventStream.Version : (long?) null,
+                cancellationToken
             );
 
             MyVersion = result.Version;
@@ -121,7 +127,7 @@ namespace Examples
             if (result.Version % 10 == 0)
                 await _es.CreateSnapshotAsync(_streamId,
                     result.Version,
-                    new SnapshotModel { Data = this },
+                    new SnapshotModel {Data = this},
                     cancellationToken: cancellationToken
                 );
 
@@ -139,12 +145,12 @@ namespace Examples
         }
 
         /// <summary>
-        /// Builds the specified fail not found.
+        ///     Builds the specified fail not found.
         /// </summary>
         /// <param name="failNotFound">if set to <c>true</c> [fail not found].</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>Task.</returns>
-        /// <exception cref="DocumentNotFoundException">StreamId {_streamId}</exception>
+        /// <exception cref="EventStoreNotFoundException">StreamId {_streamId}</exception>
         public async Task Build(bool failNotFound = false, CancellationToken cancellationToken = default)
         {
             if (!(_myEventStream is null))
@@ -155,14 +161,14 @@ namespace Examples
             if (_myEventStream is null)
             {
                 if (failNotFound)
-                    throw new DocumentNotFoundException($"StreamId {_streamId} not found");
+                    throw new EventStoreNotFoundException($"StreamId {_streamId} not found");
 
                 return;
             }
 
             if (!(_myEventStream.Snapshot is null))
             {
-                var data = (MongoAggregate)_myEventStream.Snapshot.Data;
+                var data = (MongoAggregate) _myEventStream.Snapshot.Data;
 
                 SomeValue = data.SomeValue;
                 AggregateId = data.AggregateId;